+++ src/language/syntax/parser.ts
@@ -3,21 +3,36 @@ import { MapDocumentNode, ProfileDocumentNode } from '@superfaceai/ast';
 import { SyntaxError } from '../error';
 import { Lexer } from '../lexer/lexer';
 import { Source } from '../source';
-import { RuleResult, SyntaxRule } from './rule';
+import { SyntaxRule } from './rule';
 import * as map from './rules/map';
 import { profile } from './rules/profile';
 
-function parseRuleResult<N>(
+export function parseRuleResult<N>(
   rule: SyntaxRule<N>,
   source: Source,
   skipSOF?: boolean
-): RuleResult<N> {
+): { kind: 'success'; value: N } | { kind: 'failure'; error: SyntaxError } {
   const lexer = new Lexer(source);
   if (skipSOF === true) {
     lexer.next();
   }
 
-  return rule.tryMatch(lexer);
+  const result = rule.tryMatch(lexer);
+  if (result.kind === 'match') {
+    return { kind: 'success', value: result.match };
+  } else {
+    const error = SyntaxError.fromSyntaxRuleNoMatch(source, result);
+
+    // print the formatted error on debug log level
+    if (process.env.LOG_LEVEL === 'debug') {
+      console.debug(error.format());
+    }
+
+    return {
+      kind: 'failure',
+      error,
+    };
+  }
 }
 
 /**
@@ -33,17 +48,11 @@ export function parseRule<N>(
 ): N {
   const result = parseRuleResult(rule, source, skipSOF);
 
-  if (result.kind === 'nomatch') {
-    const error = SyntaxError.fromSyntaxRuleNoMatch(source, result);
-
-    if (process.env.LOG_LEVEL === 'debug') {
-      console.debug(error.format());
-    }
-
-    throw error;
+  if (result.kind === 'failure') {
+    throw result.error;
   }
 
-  return result.match;
+  return result.value;
 }
 
 export function parseProfile(source: Source): ProfileDocumentNode {